requirements.txt
apturl==0.5.2
atomicwrites==1.1.5
attrs==24.3.0
Automat==0.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
blosc==1.7.0
Brlapi==0.7.0
cattrs==24.1.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==19.1.16
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coremltools==6.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dask==2.8.1+dfsg
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.16.1
fsspec==2024.12.0
future==0.18.2
gssapi==1.6.1
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.27.0
hyperlink==19.0.0
idna==2.8
imageio==2.4.1
imgproc==0.0.2.1
importlib-metadata==1.5.0
importlib-resources==5.4.0
incremental==16.10.1
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.0
jedi==0.15.2
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.4.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linformer==0.2.3
locket==0.2.0
lockfile==0.12.2
louis==3.12.0
lpips==0.1.4
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.1.2
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
netifaces==0.10.4
networkx==2.4
nose==1.3.7
numexpr==2.7.1
numpy==1.17.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
onnx==1.17.0
onnx-coreml==1.3
opencv-python==4.10.0.84
openpyxl==3.0.3
packaging==20.3
pandas==0.25.3
paramiko==2.6.0
parso==0.5.2
partd==1.0.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.0
prompt-toolkit==2.0.10
protobuf==3.20.1
psutil==5.5.1
py==1.8.1
pyaml==24.12.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gflags==1.5.1
pytorch-msssim==1.0.0
pytz==2019.3
PyWavelets==0.5.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
regex==2024.11.6
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
safetensors==0.4.5
scikit-image==0.16.2
scipy==1.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.3
soupsieve==1.9.5
ssh-import-id==5.10
sympy==1.13.3
systemd-python==234
tables==3.6.1
tokenizers==0.20.3
toolz==0.9.0
torch==2.4.0
torchvision==0.19.1
tqdm==4.66.6
traitlets==4.3.3
transformers==4.46.3
triton==3.0.0
Twisted==18.9.0
typing==3.7.4.3
typing-extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
zipp==3.7.0
zope.interface==4.7.1
